Extended block production test with observer node
Problem: there is no way to test catch-up works even in the case of
large blocks.

Solution: add an observer node to block production priority test.
The node will be stopped at the start and started at the very end of the
test and be checked to catch-up with the block producing node.