Use headers for v1 pubsub block topic
Receive headers on a new pubsub topic and propagate them all the way
through to catchup, as alternative to block.

Existing synchronization code relies heavily on full blocks becoming
available, hence a hack was made: when a valid and relevant header
enters catchup system, download of a block will be requested and only
when the full block is available, processing will continue.